task inputs updated

+++ Tasks/KubernetesManifestV0/src/actions/delete.ts
@@ -7,6 +7,12 @@ import * as TaskInputParameters from '../models/TaskInputParameters';
 
 export async function deleteResources() {
     let args = TaskInputParameters.args;
+
+    if (args == null || args.length == 0)
+    {
+        throw (tl.loc("ArgumentsInputNotSupplied"));
+    }
+
     let kubectl = new Kubectl(await utils.getKubectl(), TaskInputParameters.namespace);
     var result = kubectl.delete(args);
     utils.checkForErrors([result]);
